+++ src/c/partition/algo/sgl_dyn.c
@@ -24,7 +24,7 @@ void sgl_dyn(Partitions *p, bool *result)
             bool found = false;
             #pragma omp parallel default(none) shared(row, size, numOfCombinations, half_problem_sum, result, i, found)
             #pragma omp for schedule(dynamic, 10)
-            for (int j = 0; j < numOfCombinations; j++)
+            for (unsigned long long int j = 0; j < numOfCombinations; j++)
             {
                 if (found)
                     continue;
